# docs/api_contracts/openapi_v1.yaml
openapi: 3.0.0
info:
  title: Traveller Orchestrator API
  description: API contract for the Intelligent Travel Agent Platform.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server

paths:
  # --------------------
  # Auth Endpoints
  # --------------------
  /auth/login:
    post:
      summary: Login with OAuth Provider
      operationId: login
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                provider:
                  type: string
                  enum: [google, apple]
                token:
                  type: string
                  description: ID Token from the provider
      responses:
        '200':
          description: Successful Login
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'

  /auth/me:
    get:
      summary: Get Current User Profile
      operationId: getMe
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current User Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # --------------------
  # Trip Endpoints
  # --------------------
  /trips:
    post:
      summary: Create a new Trip (Plan)
      operationId: createTrip
      tags:
        - Trips
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - destination
                - startDate
                - endDate
              properties:
                destination:
                  type: string
                  example: "Paris, France"
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                budget:
                  type: number
                  description: Total budget in user's currency
                interests:
                  type: array
                  items:
                    type: string
                  example: ["Museums", "Food", "History"]
      responses:
        '201':
          description: Trip Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'

  /trips/{id}:
    get:
      summary: Get Trip Details by ID
      operationId: getTrip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      tags:
        - Trips
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Trip Details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trip'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string

    Trip:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        destination:
          type: string
        status:
          type: string
          enum: [draft, confirmed, completed]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        itinerary:
          type: object
          description: Structured itinerary data (days, activities)